+++ src/main/java/com/woof/api/member/config/PasswordEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.woof.member.config;
+package com.woof.api.member.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
