+++ src/main/java/com/woof/api/member/config/SecurityConfig.java
@@ -1,6 +1,8 @@
-package com.woof.member.config;
+package com.woof.api.member.config;
 
-import com.woof.member.service.MemberService;
+import com.woof.api.member.jwt.filter.JwtFilter;
+import com.woof.api.member.jwt.utils.TokenProvider;
+import com.woof.api.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
