+++ src/main/java/com/woof/api/member/model/requestdto/PostMemberReq.java
@@ -1,4 +1,4 @@
-package com.woof.member.model.requestdto;
+package com.woof.api.member.model.requestdto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,11 +7,11 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@NoArgsConstructor
 @AllArgsConstructor
-public class GetEmailConfirmReq {
+@NoArgsConstructor
+public class PostMemberReq {
 
     private String email;
-    private String jwt;
-    private String uuid;
+    private String password;
+    private String nickname;
 }
